fix(auth-code-flow): omit scope when requesting access token
Providing `scope` as a body parameter to the access token endpoint
will lead to an error on the Zalando oAuth2 provider. The problem
is fixed via ommiting the `scope` parameter.

The `scope` parameter does not need to be defined since the
`AUTHORIZATION_CODE_GRANT` only supports the `/employees` realm
which has only one default scope (`uid`).

see #28