feat(api): optional query parameters
When requesting an access token (via `TokenCache` or via
`getAccessToken`) one can now specify an object `queryParams`
in the `OAuthConfig`. The given key-value pairs will be added
to the reuqest as query parameters.

`createAuthCodeRequestUri` accepts an optional parameter
`queryParams`. The given key-value pairs will be added to the
returned url string as query parameters.

close #39